Implement webhooks; change job and entity policies; fix tests

+++ app/controllers/beach_api_core/v1/jobs_controller.rb
@@ -19,10 +19,12 @@ module BeachApiCore
     end
 
     def show
+      authorize @job
       render_json_success(@job, :ok, root: :job)
     end
 
     def destroy
+      authorize @job
       if @job.destroy
         render_json_success(@job, :ok, root: :job)
       else
